+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - Used newly introduced trigger engine feature to delay the `Enable All Triggers` dialogue till it can be properly called (requires `1.22.1`)
   - `Trove Attack: Handler` added special case to handle base `Unarmed Attack`
   - Updated `Test Animations` to actually run properly
+  - Added `trigger-engine` and `sequencer` as explicit dependencies
 
 ## 0.9.3